const_serialize

Trait SerializeConst

pub unsafe trait SerializeConst: Sized {
    const MEMORY_LAYOUT: Layout;
    const _ASSERT: () = _;
}

A trait for types that can be serialized and deserialized in const.

Safety

The layout must accurately describe the memory layout of the type

Required Associated Constants

const MEMORY_LAYOUT: Layout

The memory layout of the type. This type must have plain old data; no pointers or references.

Provided Associated Constants

const _ASSERT: () = _

Assert that the memory layout of the type is the same as the size of the type

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.